Update the reference frames for NO_TIME_DELAY
We need a particular `t` value to use when computing the position
of a vertex that we are rendering. For the realistic model, we set
`t` to be the time taken for the light to reach us from the vertex,
which is just c * distance to the vertex.

For the `NO_TIME_DELAY` case, we choose a reference frame
for which we consider all `t`'s to be emitted at the same time and
reach the camera at the same time (since we ignore the light
travel time.) We have two sensible cases: using the world as the
rest frame, or the camera as the rest frame.